### 2020-07-10 #2467 @whedon
[PRE REVIEW]: causal-curve: A Python Causal Inference Package to Estimate Causal Dose-Response Curves
**Submitting author:** @ronikobrosly (<a href="http://orcid.org/0000-0003-0363-9662">Roni Kobrosly</a>)
**Repository:** <a href="https://github.com/ronikobrosly/causal-curve" target ="_blank">https://github.com/ronikobrosly/causal-curve</a>
**Version:** v0.3.2
