### 2020-07-18 48fad65
small edits to paper, bib, and readme
