+++ paper/paper.md
@@ -7,7 +7,7 @@ tags:
   - machine learning
 
 authors:
-  - name: Roni W. Kobrosly^[Custom footnotes for e.g. denoting who the corresspoinding author is can be included like this.]
+  - name: Roni W. Kobrosly
     orcid: 0000-0003-0363-9662
     affiliation: "1, 2" # (Multiple affiliations must be quoted)
 affiliations:
@@ -30,9 +30,9 @@ from observational rather than experimental data, correcting for the biases that
 to analyzing observational data (e.g. confounding and selection bias) [@Hernán:2020].
 
 Although significant research and implementation effort has gone towards methods in
-causal inference to estimate the effects of binary treatments (e.g. did the population receive
-treatment "A" or "B"), much less has gone towards estimating the effects of continuous treatments.
-This is unfortunate because there are a large number of use cases in research
+causal inference to estimate the effects of binary treatments (e.g. what was the effect of
+treatment "A" or "B"?), much less has gone towards estimating the effects of continuous treatments.
+This is unfortunate because there are a great number of inquiries in research
 and industry that could benefit from tools to estimate the effect of
 continuous treatments, such as estimating how:
 
@@ -71,13 +71,12 @@ coverage is monitored via codecov and is presently above 90%.
 
 The `GPS` method was originally described by Hirano [@Hirano:2004],
 and expanded by Moodie [@Moodie:2010] and more recently by Galagate [@Galagate:2016]. GPS is
-an extension of the standard propensity tool method. It is the treatment assignment density calculated
+an extension of the standard propensity tool method and is essentially the treatment assignment density calculated
 at a particular treatment (and covariate) value. Similar to the standard propensity score approach,
 the GPS random variable is used to balance covariates. At the core of this tool, generalized linear
-models are used to estimate the GPS, and generalized additive models are used to estimate the
-final causal curve. Compared with the package’s TMLE method,
-this GPS method is more computationally efficient, better suited for large datasets,
-but produces significantly wider confidence intervals.
+models are used to estimate the GPS, and generalized additive models are used to estimate the smoothed
+final causal curve. Compared with this package’s TMLE method, this GPS method is more
+computationally efficient, better suited for large datasets, but produces significantly wider confidence intervals.
 
 
 ![Example of a causal curve generated by the GPS tool.\label{fig:example}](welcome_plot.png)
@@ -96,8 +95,8 @@ against model misspecification, incorporates more powerful machine learning tech
 smaller confidence intervals, however it is less computationally efficient.
 
 `causal-curve` allows for continuous mediation assessment with the `Mediation` tool. As described
-by Imai this approach provides a general approach to mediation analysis that invokes the  
-potential outcomes / counterfactual framework [@Imai:2010]. While this approach can handle a
+by Imai this approach provides a general approach to mediation analysis that invokes the potential
+outcomes / counterfactual framework [@Imai:2010]. While this approach can handle a
 continuous mediator and outcome, as put forward by Imai it only allows for a binary treatment. As
 mentioned above with the `TMLE` approach, the tool creates a series of binary treatment comparisons
 and connects them to show the user how mediation varies as a function of the treatment. An interpretable,
