+++ paper/paper.bib
@@ -35,12 +35,13 @@
     doi = {10.2202/1557-4679.1181},
 }
 
-@article{van_der_Laan:2006,
-    author = {{van der Laan}, M. and {Rubin}, D.},
-    title = "{Targeted maximum likelihood learning.}",
-    journal = {The International Journal of Biostatistics},
-    year = 2006,
-    volume = 2,
+@misc{van_der_Laan:2006,
+  author = {{van der Laan}, M. and {Rubin}, D.},
+  title = "{Targeted maximum likelihood learning.}",
+  year = {2006},
+  publisher = {University of California, Berkeley},
+  journal = {​U.C. Berkeley Division of Biostatistics Working Paper Series},
+  url = {​https://biostats.bepress.com/ucbbiostat/paper213/}
 }
 
 @article{Imai:2010,
