+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2020, Roni Kobrosly'
 author = 'Roni Kobrosly'
 
 # The full version, including alpha/beta/rc tags
-release = '0.3.2'
+release = '0.3.3'
 
 # -- General configuration ---------------------------------------------------
 
