+++ paper/paper.bib
@@ -41,7 +41,7 @@
   year = {2006},
   publisher = {University of California, Berkeley},
   journal = {​U.C. Berkeley Division of Biostatistics Working Paper Series},
-  url = {​https://biostats.bepress.com/ucbbiostat/paper213/}
+  url = {​https://biostats.bepress.com/ucbbiostat/paper213}
 }
 
 @article{Imai:2010,
