### 2020-07-21 #2467 @oliviaguest
@ronikobrosly when you get a sec, please add the missing DOI and then recompile the PDF using the `@whedon generate pdf` command. ☺️
