+++ paper/paper.bib
@@ -26,7 +26,7 @@
   	Year = 2004
 }
 
-@article{van_der_Laan:2010,
+@article{van_der_Laan:2010a,
     author = {{van der Laan}, M. and {Gruber}, S.},
     title = "{Collaborative double robust penalized targeted maximum likelihood estimation.}",
     journal = {The International Journal of Biostatistics},
@@ -35,12 +35,13 @@
     doi = {10.2202/1557-4679.1181},
 }
 
-@misc{van_der_Laan:2006,
-  author = {{van der Laan}, M. and {Rubin}, D.},
-  title = "{Targeted maximum likelihood learning.}",
-  year = {2006},
-  publisher = {University of California, Berkeley},
-  journal = {​U.C. Berkeley Division of Biostatistics Working Paper Series},
+@article{van_der_Laan:2010b,
+    author = {{van der Laan}, M.},
+    title = "{Targeted Maximum Likelihood Based Causal Inference: Part I.}",
+    journal = {The International Journal of Biostatistics},
+    year = 2010,
+    volume = 6,
+    doi = {10.2202/1557-4679.1211},
 }
 
 @article{Imai:2010,
