+++ paper/paper.bib
@@ -26,7 +26,7 @@
   	Year = 2004
 }
 
-@article{van_der_Laan:2010a,
+@article{van_der_Laan:2010,
     author = {{van der Laan}, M. and {Gruber}, S.},
     title = "{Collaborative double robust penalized targeted maximum likelihood estimation.}",
     journal = {The International Journal of Biostatistics},
@@ -35,7 +35,7 @@
     doi = {10.2202/1557-4679.1181},
 }
 
-@article{van_der_Laan:2010b,
+@article{van_der_Laan:2010,
     author = {{van der Laan}, M.},
     title = "{Targeted Maximum Likelihood Based Causal Inference: Part I.}",
     journal = {The International Journal of Biostatistics},
