+++ paper/paper.bib
@@ -35,15 +35,6 @@
     doi = {10.2202/1557-4679.1181},
 }
 
-@article{van_der_Laan:2010,
-    author = {{van der Laan}, M.},
-    title = "{Targeted Maximum Likelihood Based Causal Inference: Part I.}",
-    journal = {The International Journal of Biostatistics},
-    year = 2010,
-    volume = 6,
-    doi = {10.2202/1557-4679.1211},
-}
-
 @article{Imai:2010,
     author = {{Imai}, K., {Keele}, L., and {Tingley}, D.},
     title = "{A General Approach to Causal Mediation Analysis.}",
