### 2020-07-21 #2467 @ronikobrosly
@oliviaguest , because it was a Berkeley internal dept paper and it seems like the other TMLE reference covers it I ended up removing that reference.
