### 2020-07-31 #2523 @tomfaulkenberry
I have completed my review, and I ran into no issues.  The software paper clearly describes the audience and intended use of the software, and the software installed perfectly fine on my machine (2019 Macbook Air running Catalina, Python version 3.8 installed via Homebrew).  All tutorials in the documentation were clear and I was able to recreate the examples on my own).

Overall, I can conclude my review now with a positive endorsement...this is a very nice piece of software and is perfectly within the scope of JOSS.

Thanks for asking me to review.

Tom Faulkenberry
Tarleton State University
