### 2020-08-03 #2523 @alexjonesphd
Hi all,

I installed the package fine via pip on a clean conda environment (3.7; Mac OS Mojave), got all the dependencies, but was seeing some import errors surrounding XGBoost, possibly due to OpenMP runtime. I've raised an issue on the code repo -[https://github.com/ronikobrosly/causal-curve/issues/15].

I've no experience of working with homebrew but my own poking around leads me to believe OpenMP which XGBoost needs may require install via that, which is a tricky step to get the package running.
