### 2020-08-03 #2523 @ronikobrosly
Hi @alexjonesphd , thanks for raising this. The XGBoost package does require OpenMP: https://xgboost.readthedocs.io/en/latest/build.html#building-on-osx

I'm of two minds on this. On one hand XGBoost is a really nice ML package that is pretty widely used, but on the other hand maybe better to stick to a ML approach here that doesn't have that requirement you raised. I guess there are a couple options for this:

1) I could leave XGBoost in for now and specifically mention the OpenMP requirement and any other requirements for XGBoost in the documentation. Right now I don't mention these XGBoost requirements.
2) I could replace XGBoost with another package's gradient boosting implementation. Maybe I could use `sklearn`'s implementation.

What do you think?
