### 2020-08-03 #2523 @alexjonesphd
Hey @ronikobrosly, no problem, sorry to raise the issue, the package looks great and was excited to play with it!

My own take would be as follows, though feel free to disagree:
 - I think that the use of a less tricky to install gradient boosting implementation would make the package more widely accessible to the community. Typically I think most users would want something that installs easily without additional work with something like homebrew. This has always been the advantage of R in that things are (usually!) easier to install and get off the ground. I say this as a psychologist who uses a lot of ML approaches but would much rather work in Python, but I'm not savvy enough to do lots of extra steps to get things installed. I'm probably a typical use case so would rather a simpler installation step, assuming switching from XGBoost to the `sklearn` class wouldn't hurt the performance of the package too much! I just feel that extra accessibility would be a huge bonus.
