### 2020-08-03 #2523 @ronikobrosly
It's no problem at all to raise this @alexjonesphd . I hear you. I'll go ahead and try swapping out XGBoost for `sklearn`'s boosting algorithm.

Is this completely blocking your installation and the rest of your review? If it is, I'll try to get to this ASAP (but it might be next weekend).

If you're able to install this and complete the review, is it okay if I hold off on revisions until I get the other reviews back? Just so I can tackle other comments all at once.
