### 2020-08-03 #2523 @alexjonesphd
Afraid so @ronikobrosly - I got the XGBoost error on line 1, e.g.

`from causal_curve import GPS` throws the error. I don't know how much the package relies on XGBoost, I can tweak the code to prevent it importing and review the materials if required. Otherwise, happy to wait, and maybe @cmparlettpelleriti has other points!
