### 2020-08-03 2cf78c4
updated comments and docs
