### 2020-08-03 ba94fe1
Feature/swap out xgboost (#16)
* removed xgboost from TMLE

* updated comments and docs

* updated more docs
