+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r") as fh:
 
 setuptools.setup(
     name="causal-curve",
-    version="0.3.3",
+    version="0.3.4",
     author="Roni Kobrosly",
     author_email="roni.kobrosly@gmail.com",
     description="A python library with tools to perform causal inference using \
@@ -38,7 +38,6 @@ setuptools.setup(
         'scikit-learn',
         'scipy',
         'six',
-        'statsmodels',
-        'xgboost'
+        'statsmodels'
     ]
 )
