### 2020-08-03 ronikobrosly/causal-curve#15 @ronikobrosly
Just pushed a new version (0.3.4) that doesn't use XGBoost. So I think this should be solved.
