### 2020-08-03 #2523 @ronikobrosly
@alexjonesphd , I just merged a PR that removes XGBoost as a dependency. We're now using `sklearn`'s `GradientBoostingClassifier` and `GradientBoostingRegressor`. It's a tad slower but the results look basically the same. It should also be easier to install this now 👍 👍

Feel free to pip install the latest version (now it's 0.3.4)
