### 2020-08-03 #2523 @alexjonesphd
That was fast @ronikobrosly! The no-rest-until-code-is-fixed bug, I know it well! Just clean installed it and importing `GPS` worked fine. I'll go through the examples tomorrow (in the UK!). Thanks for fixing so quickly, though sorry to hear its a little slower!
