### 2020-08-04 #2523 @alexjonesphd
Thanks again for editing the codebase so quickly. I've now had a chance to work with the package and it seems great. The API is nicely modelled on `sklearn` so the instantiating a class and calling `fit` was second nature. All the examples worked well for me, I was able to reproduce them easily.

Perhaps my only comment would be expansion of the software paper or documentation to make it a little more accessible to those who don't know causal inference and perhaps why they would want to use the package. I'll freely admit I don't know fully what is being computed here but the codebase works well and is easy to use. It may be that this is beyond the scope of the journal and that's fine, but a little introduction to causal inference and why the package can help with that might make it more accessible. Otherwise, this is great!
