### 2020-08-04 #2523 @ronikobrosly
Thanks for the kind words @alexjonesphd ! Sure thing, I can add a bit more text in the paper and/or documentation about causal inference and why they would want to use it. I'll add this in after I hear back from @cmparlettpelleriti .
