### 2020-08-10 #2523 @cmparlettpelleriti
hello! Sorry for the delay, I'm slowly working my way through the checklist 🙂

Get Outlook for iOS<https://aka.ms/o0ukef>
