### 2020-08-14 93f9219
doc revisions (#17)
