### 2020-08-14 #2523 @ronikobrosly
Hi @alexjonesphd , I went ahead and revised the [online documentation](https://causal-curve.readthedocs.io/en/latest/). There is now an `introduction` section and I tried to make it a gentle intro to experimentation, interpreting the causal curve, etc.
