### 2020-08-14 #2523 @alexjonesphd
Thanks @ronikobrosly that looks great. Nice work! As before very happy to endorse and hope this gets some serious use by the Python community.
