### 2020-08-17 #2523 @cmparlettpelleriti
Thanks for your patience waiting for my review, I thought this was a beautifully organized paper and repository. As someone who doesn't do a ton of causal modeling, I still found it relatively easy to follow. You got me excited to do more causal modeling!

Some minor fixes:

**In the main example file:** Under Distributions of Scaled Test Scores the x axis labels might be incorrect (they still say blood lead, but I think they’re distributions of test scores)
I got an error when running the code out of the box. I've attached a screenshot here. It's likely my fault, but after installing using pip, downloading the data and running the entire notebook, I got the error so I imagine others could too.

**In Mediation_example.rst** it should be “developed by Imai et al.”

I really enjoyed your comment about mirroring sklearn syntax, I think that's smart and will reduce cognitive load for people who are coming to this from other ML applications. My one larger suggestion is optional, but as someone who doesn't deeply know all the concepts you're implementing, linking out to other resources that more completely explain the math of the models would be so useful. If there are any open source books, or great blogs you can include here, I think that would help new learners a lot.

This is super cool, I very much enjoyed reviewing it!
