### 2020-08-17 #2523 @ronikobrosly
Thanks so much for the review and the kind words @cmparlettpelleriti 😄 ! Ok I'll look into your comments ASAP and let you know here how I fixed them.
