### 2020-08-17 #2523 @ronikobrosly
@cmparlettpelleriti , quick question for you: Would you mind telling me what version of `scipy` you're using? Maybe the easiest way to find out would be to type `pip freeze` in the terminal within your virtual environment.
