### 2020-08-17 #2523 @ronikobrosly
The reason I'm asking @cmparlettpelleriti is I think that older versions of `scipy` give that error. I created a fresh virtual environment, pip installed requirements (including the latest `scipy`), ran notebook, and didn't see the error. I'm running version `1.5.2`.

I'd rather not pin the versions of the dependencies in the `requirements.txt` file but maybe what I could do is specify in the documentation that scipy >= 1.5.2 is required? What do you think?
