### 2020-08-18 #2523 @cmparlettpelleriti
@ronikobrosly great suggestion! I was running 1.4.1, I upgraded to 1.5.2 and now got this error. Perhaps a different package that I need to update (I'm not on my normal work computer so things may be out of date; full list copied below)? Sorry about this, it seems like this is more user error than a problem with your code.

I totally agree that you don't want to specify a specific version, I think your idea is wonderful but I'll leave the final decision up to you, I think you know better what your users will need.

absl-py==0.9.0
alabaster==0.7.12
appdirs==1.4.4
appnope==0.1.0
arabic-reshaper==2.0.15
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
black==19.10b0
bleach==3.1.1
causal-curve==0.3.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coverage==5.2.1
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
descartes==1.1.0
docutils==0.16
dukpy==0.2.2
entrypoints==0.3
esprima==4.0.1
et-xmlfile==1.0.1
freetype-py==2.1.0.post1
future==0.18.2
gast==0.3.3
gevent==20.4.0
gitdb==4.0.4
GitPython==3.1.1
glfw==1.11.0
greenlet==0.4.15
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
imagesize==1.2.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
javascripthon==0.11
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json-tricks==3.15.2
json5==0.9.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
macropy3==1.1.0b2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mizani==0.6.0
more-itertools==8.4.0
moviepy==1.0.2
msgpack==1.0.0
msgpack-numpy==0.4.5
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.1
numpydoc==1.1.0
opencv-python==4.2.0.34
openpyxl==3.0.3
packaging==20.4
palettable==3.3.0
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pathspec==0.8.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotnine==0.6.0
pluggy==0.13.1
proglog==0.1.9
progressbar2==3.51.4
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.7.0
PsychoPy==2020.1.2
ptyprocess==0.6.0
py==1.9.0
pycparser==2.20
pygam==0.8.0
pyglet==1.5.4
Pygments==2.5.2
pyobjc-core==6.2
pyobjc-framework-Cocoa==6.2
pyobjc-framework-Quartz==6.2
PyOpenGL==3.1.5
pyOpenSSL==19.1.0
pyosf==1.0.5
pyparsing==2.4.6
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyrsistent==0.15.7
pyserial==3.4
pytest==5.4.3
python-bidi==0.4.2
python-dateutil==2.8.1
python-gitlab==2.2.0
python-utils==2.4.0
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
questplus==2019.4
regex==2020.7.14
requests==2.23.0
scikit-learn==0.22.1
scipy==1.5.2
seaborn==0.10.1
Send2Trash==1.5.0
six==1.14.0
smmap==3.0.2
snowballstemmer==2.0.0
sounddevice==0.3.15
SoundFile==0.10.3.post1
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
statsmodels==0.11.1
tables==3.6.1
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.3
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
wxPython==4.1.0
xarray==0.15.1
xgboost==1.1.1
xlrd==1.2.0
