### 2020-08-18 #2523 @ronikobrosly
@cmparlettpelleriti whoops I forgot to mention this yesterday: I actually got this `Expect x to not have duplicates` too yesterday. This is because of a real bug in the code (this line needs to be fixed: https://github.com/ronikobrosly/causal-curve/blob/93f9219b41e3174a14b96a0ebb271da730b92c56/causal_curve/mediation.py#L520). I fixed it locally and it seems to work but I haven't pushed anything to the repo yet. I'll let you know when I finish up the rest of the changes you asked for (should be soon).

So basically, you uncovered a bug 🙃
