### 2020-08-20 #2523 @ronikobrosly
Ok @cmparlettpelleriti , I just pushed those changes. If you upgrade your version of `causal-curve` to version `0.3.7`, the notebook should work for you all the way through. Could you verify that for you me?

Also, I made those misc changes:
- I fixed the axis titles in the plots in the notebook and documentation
- Fixed the Imai typo
- I add a couple of links to useful resources in the introduction section of the documentation.
