### 2020-08-20 #2523 @cmparlettpelleriti
@ronikobrosly Everything looks wonderful, and the notebook now works perfectly after updating to 0.3.7 !
