### 2020-08-31 #2523 @kthyng
Hi @ronikobrosly! I see @oliviaguest set the version as v0.3.2 but your repo has a version v0.3.7 — which is correct for the JOSS archive/publication?
