### 2020-08-31 #2523 @kthyng
> Hi @oliviaguest I just changed the zenodo repo title like you asked 👍

Am I seeing an old version of the Zenodo archive? When I look at it, I see the title still needing to change. It should be "causal-curve: A Python Causal Inference Package to Estimate Causal Dose-Response Curves".
