### 2020-08-31 #2523 @kthyng
The paper looks good! Just read through.
