### 2020-08-31 #2523 @ronikobrosly
Hi @kthyng , nice to meet you! You're right, I changed the zenodo repo name but forgot to publish it. It should be right now. Could we use v0.3.7 for publication?
