### 2020-08-31 #2523 @ronikobrosly
> Congrats to @ronikobrosly on your new publication!! Thanks to editor @oliviaguest and to reviewers @cmparlettpelleriti, @tomfaulkenberry, and @alexjonesphd — we rely on your expertise and hard work to make this process happen.
>
> I'll close this issue once the doi resolves. Thanks all!!

Thanks so much @oliviaguest, @cmparlettpelleriti, @tomfaulkenberry, @alexjonesphd, and @kthyng !
